==> 1.1-data_discovery.ipynb <==
#import libraries
import pandas as pd
import numpy as np


#download first file to discover how it looks
df = pd.read_csv("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/2014/Data_raw_punctuality_201401.csv")


#count how much data we have
df.count()

#look at first 100 rows
df.tail(50)

#get final destination by grouping per train no. and date of departure
df['destination'] = df.groupby(['DATDEP', 'TRAIN_NO'], sort=False)['PTCAR_LG_NM_NL'].transform('last')

#get departure station by grouping per train no. and date of departure
df['departure'] = df.groupby(['DATDEP', 'TRAIN_NO'], sort=False)['PTCAR_LG_NM_NL'].transform('first')

df.head(1220)


#get only rows where the train departed from BRUSSEL-CENTRAAL
df2 = df[df.PTCAR_LG_NM_NL == 'BRUSSEL-CENTRAAL']


#add column to define if train was delayed or not
df2['delayed'] = np.where(df2['DELAY_DEP'] > 360, True, False)


df2.count()
