==> 1.2-data_train_delays.ipynb <==
#import libraries
import pandas as pd
import numpy as np

# apply same rules to all trian delays files
months = ["01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12"]
df_train_delays = pd.DataFrame(columns=('DATDEP', 'TRAIN_NO', 'RELATION', 'TRAIN_SERV', 'PTCAR_NO',
       'LINE_NO_DEP', 'REAL_TIME_ARR', 'REAL_TIME_DEP', 'PLANNED_TIME_ARR',
       'PLANNED_TIME_DEP', 'DELAY_ARR', 'DELAY_DEP', 'RELATION_DIRECTION',
       'PTCAR_LG_NM_NL', 'LINE_NO_ARR', 'PLANNED_DATE_ARR', 'PLANNED_DATE_DEP',
       'REAL_DATE_ARR', 'REAL_DATE_DEP', 'departure', 'destination', 'delayed'))

for i in range(9):
       for j in range(12):
              df = pd.read_csv("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/20" + str(14+i) + "/Data_raw_punctuality_20" + str(14+i) + months[j] + ".csv")
              df['PTCAR_LG_NM_NL'][pd.isna(df.REAL_TIME_DEP)]
              df['departure'] = df.groupby(['DATDEP', 'TRAIN_NO'], sort=False)['PTCAR_LG_NM_NL'].transform('first')
              df['destination'] = df.groupby(['DATDEP', 'TRAIN_NO'], sort=False)['PTCAR_LG_NM_NL'].transform('last')
              df2 = df[df.PTCAR_LG_NM_NL == 'BRUSSEL-CENTRAAL']
              df2['delayed'] = np.where(df2['DELAY_DEP'] > 360, True, False)
              df_train_delays = df_train_delays.append(df2)


# count the final rows
df_train_delays.count()

# add year, month, day columns
df_train_delays["year"] = df_train_delays["DATDEP"].apply(lambda x: x[5:])
df_train_delays["month"] = df_train_delays["DATDEP"].apply(lambda x: x[2:5])
df_train_delays["day"] = df_train_delays["DATDEP"].apply(lambda x: x[:2])


df_train_delays["month"].unique()


# change month to month number
df_train_delays["month"] = df_train_delays["month"].map({"JAN": "01", "FEB": "02", "MAR": "03", "APR": "04", "MAY": "05", "JUN": "06", "JUL": "07", "AUG": "08", "SEP": "09", "OCT": "10", "NOV": "11", "DEC": "12"})


df_train_delays.head()


# get the hour from the planned departure
df_train_delays["hour_time_dep"] = df_train_delays["PLANNED_TIME_DEP"].astype("datetime64").dt.round("H").dt.hour


#export the distinct destinations in order to get the regions
pd.DataFrame(df_train_delays['destination'].unique()).to_excel("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/train_stations.xlsx", index=False)


#export the distinct departures in order to get the regions
pd.DataFrame(df_train_delays['departure'].unique()).to_excel("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/train_stations_dep.xlsx", index=False)


# import the train stations with regions
df_regions = pd.read_excel("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/train_stations_regions.xlsx")


# merge train delays with regions from destinations and drop train station column from regions df
final_df = pd.merge(df_train_delays, df_regions, how="left", left_on = ["destination"], right_on = ["train_stations" ])

final_df = final_df.drop(columns=["train_stations"])

final_df.rename(columns={"region": "destination_region"}, inplace=True)


# merge train delays with regions from departures and drop train station column from regions df
final_df_2 = pd.merge(final_df, df_regions, how="left", left_on = ["departure"], right_on = ["train_stations" ])

final_df_2 = final_df_2.drop(columns=["train_stations"])

final_df_2.rename(columns={"region": "departure_region"}, inplace=True)




final_df_2.columns

# get weather data
df_weather = pd.read_pickle("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/weather_data.pkl")

final_df_2.dtypes

# change year, month, and day data type to integer
final_df_2["year"] = pd.to_numeric(final_df_2["year"])
final_df_2["month"] = pd.to_numeric(final_df_2["month"])
final_df_2["day"] = pd.to_numeric(final_df_2["day"])

# merge train delays with weather data and drop duplicated columns
final_df_3 = pd.merge(final_df_2, df_weather, how="left", left_on = ["year", "month", "day", "hour_time_dep"], right_on = ["year", "month", "day", "time"])
final_df_3.drop(columns=["time", "DATDEP"], inplace=True)

final_df_3.head()

# export to pickle
final_df_3.to_pickle("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/final_df.pkl")
