==> 1.3-data_weather.ipynb <==
#import libraries
import pandas as pd
import numpy as np

# read file
df = pd.read_csv("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/brussels_weather.csv")


df.head()


# drop first 9 rows
df.drop(df.index[:9], inplace=True)

# rename columns
df = df.rename(columns={"location": "date_time" ,"Grand Place Brussels": "temperature", "Grand Place Brussels.1": "relative_humidity", "Grand Place Brussels.2": "wind_speed", "Grand Place Brussels.3": "wind_direction"})

#get the year, month, day and time in separate columns
df["year"] = df["date_time"].apply(lambda x: x[:4])
df["month"] = df["date_time"].apply(lambda x: x[4:6])
df["day"] = df["date_time"].apply(lambda x: x[6:8])
df["time"] = df["date_time"].apply(lambda x: x[9:])

df.head()

# check the data types
df.dtypes

# change temperature, relative humidity, wind speed, wind direction, and time data type to numeric
df["temperature"] = pd.to_numeric(df["temperature"])
df["relative_humidity"] = pd.to_numeric(df["relative_humidity"])
df["wind_speed"] = pd.to_numeric(df["wind_speed"])
df["wind_direction"] = pd.to_numeric(df["wind_direction"])
df["year"] = pd.to_numeric(df["year"])
df["month"] = pd.to_numeric(df["month"])
df["day"] = pd.to_numeric(df["day"])

# change time data type to float
df["time"] = df["time"].apply(lambda x: x[:2]).astype(float)


df.dtypes

# drop datetime column
df.drop(columns=["date_time"], inplace=True)

df.head()

# export to pickle
df.to_pickle("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/weather_data.pkl")
