==> 2.1-data_preparation.ipynb <==
#import libraries
import pandas as pd
import numpy as np
pd.set_option('display.expand_frame_repr', False)

# import file
df = pd.read_pickle("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/final_df.pkl")

df.head()

# drop columns that will not be used for analysis

df = df.drop(columns=["TRAIN_NO", "RELATION", "PTCAR_NO", "LINE_NO_DEP"
                    , "RELATION_DIRECTION", "LINE_NO_ARR", "PLANNED_DATE_ARR"
                    , "PLANNED_DATE_DEP", "REAL_DATE_ARR", "REAL_DATE_DEP"
                    ,"PTCAR_LG_NM_NL"])

df.head()

#check which columns have nulls

df.describe(include="all")

# create a list of our conditions
conditions = [
    (df["hour_time_dep"] < 6),
    (df["hour_time_dep"] >= 6) &  (df["hour_time_dep"] < 12),
    (df["hour_time_dep"] >= 12) &  (df["hour_time_dep"] < 18),
    (df["hour_time_dep"] >= 18),
    ]

# create a list of the values we want to assign for each condition
values = ["night", "morning", "afternoon", "night"]

# create a new column and use np.select to assign values to it using our lists as arguments
df["time_of_day"] = np.select(conditions, values)

# display updated DataFrame
df.head()

# set option for pandas to display all columns at once
pd.set_option('display.max_columns', 500)


# analyze rows where the temperature is null
df[df["temperature"].isnull()]

# delete all rows where the temperature is null because these rows are 
# trains which directly departed from BRUSSEL-CENTRAAL
df.dropna(subset=["temperature"], inplace=True)

df.count()

# drop where the planned_time_arr is null
df.dropna(subset=["PLANNED_TIME_ARR"], inplace=True)

df.count()

# drop columns that will not be used
df = df.drop(columns=["REAL_TIME_ARR", "REAL_TIME_DEP", "PLANNED_TIME_ARR"
                    , "PLANNED_TIME_DEP", "DELAY_ARR", "departure", "destination"
                    , "hour_time_dep"])

df.head()

# check dtypes
df.dtypes

df.describe(include="all")

# change data types
df["TRAIN_SERV"] = df["TRAIN_SERV"].astype(str)
df["DELAY_DEP"] = pd.to_numeric(df["DELAY_DEP"])
df["delayed"] = df["delayed"].astype(bool)
df["destination_region"] = df["destination_region"].astype(str)
df["departure_region"] = df["departure_region"].astype(str)
df["time_of_day"] = df["time_of_day"].astype(str)

df.dtypes

# rename columns
df.rename(columns={"TRAIN_SERV":"train_serv", "DELAY_DEP":"delay_dep"}, inplace=True)

df.head()

# create a date column
df["date"] = df["day"].astype(str) + "-" + df["month"].astype(str) + "-" + df["year"].astype(str)

df.head()

# change data type to date for the date column
df["date"] = df["date"].apply(pd.to_datetime)

# check final dtypes
df.dtypes

df.tail()

# change month column to string
df["month"] = df["month"].map({1: "JAN", 2: "FEB", 3: "MAR", 4: "APR"
                , 5: "MAY", 6: "JUN", 7: "JUL", 8: "AUG", 9: "SEP"
                , 10: "OCT", 11: "NOV", 12: "DEC"})

df.head()

# drop year, day and delay_dep columns
df.drop(columns=["year", "day", "delay_dep"], inplace=True)

df.head()

# add column for day of the week
df['day_of_week'] = df['date'].dt.day_name()

df.head()

# final count
df.count()

# final data types
df.dtypes

#export to pickle
df.to_pickle("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/df_train_delays.pkl")
