==> 3.1-train_test_split.ipynb <==
# import libraries

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

# read df

df = pd.read_pickle("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/df_train_delays.pkl")

# get variables

df.columns

# identify y and X

y = df["delayed"]
X = df[['train_serv', 'month', 'destination_region',
       'departure_region', 'temperature', 'relative_humidity', 'wind_speed',
       'wind_direction', 'time_of_day', 'date', 'day_of_week']]

# split the data

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=3, test_size=0.25)

X_train.count()

X_test.count()

#export to pickle
X_train.to_pickle("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/train_test_data/X_train.pkl")
X_test.to_pickle("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/train_test_data/X_test.pkl")
y_train.to_pickle("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/train_test_data/y_train.pkl")
y_test.to_pickle("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/train_test_data/y_test.pkl")
