==> 5.1-one_hot_encode_normalization.ipynb <==
import numpy as np
import pandas as pd
from sklearn import preprocessing

#read pickle
X_train = pd.read_pickle("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/train_test_data/X_train_over.pkl")

X_train.columns

one_hot_encoded_data = pd.get_dummies(X_train, columns = ['train_serv', 'month'
        , 'destination_region', 'departure_region',
       'time_of_day', 'day_of_week'])


one_hot_encoded_data.head()

df_to_normalize = one_hot_encoded_data[["temperature","relative_humidity"
                                           ,"wind_speed", "wind_direction"]]

min_max = preprocessing.MinMaxScaler()
scaled_df = min_max.fit_transform(df_to_normalize.values)
normalized_df = pd.DataFrame(scaled_df, columns=["temperature","relative_humidity"
                                           ,"wind_speed", "wind_direction"])


one_hot_encoded_data["temperature"] = normalized_df["temperature"]
one_hot_encoded_data["relative_humidity"] = normalized_df["relative_humidity"]
one_hot_encoded_data["wind_speed"] = normalized_df["wind_speed"]
one_hot_encoded_data["wind_direction"] = normalized_df["wind_direction"]

one_hot_encoded_data.head()

#export to pickle
#one_hot_encoded_data.to_pickle("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/train_test_data/X_train_processed.pkl")

Test data preprocessing

#read pickle
X_test = pd.read_pickle("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/train_test_data/X_test.pkl")

X_test.reset_index(inplace=True)
X_test.head()

one_hot_encoded_data_test = pd.get_dummies(X_test, columns = ['train_serv', 'month'
        , 'destination_region', 'departure_region',
       'time_of_day', 'day_of_week'])

one_hot_encoded_data_test.head()

df_to_normalize_test = one_hot_encoded_data_test[["temperature","relative_humidity"
                                           ,"wind_speed", "wind_direction"]]

min_max = preprocessing.MinMaxScaler()
scaled_df = min_max.fit_transform(df_to_normalize_test.values)
normalized_df_test = pd.DataFrame(scaled_df, columns=["temperature","relative_humidity"
                                           ,"wind_speed", "wind_direction"])


normalized_df_test.head()

one_hot_encoded_data_test["temperature"] = normalized_df_test["temperature"]
one_hot_encoded_data_test["relative_humidity"] = normalized_df_test["relative_humidity"]
one_hot_encoded_data_test["wind_speed"] = normalized_df_test["wind_speed"]
one_hot_encoded_data_test["wind_direction"] = normalized_df_test["wind_direction"]

one_hot_encoded_data_test.head()

one_hot_encoded_data_test.drop(columns=["index"], inplace=True)
one_hot_encoded_data_test.head()

#export to pickle
one_hot_encoded_data_test.to_pickle("/Users/ilseeb/Desktop/Postgraduate/Train Punctuality Data/train_test_data/X_test_processed.pkl")
